Move recording details to new `recorder` module
